# frontend/src/lib/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Trip Planner API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/runs": {
      "get": {
        "tags": [
          "automation"
        ],
        "summary": "List Runs",
        "operationId": "list_runs",
        "parameters": [
          {
            "name": "trip_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Trip Id"
            }
          },
          {
            "name": "routine_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Routine Id"
            }
          },
          {
            "name": "kind",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "flight_api",
                    "flight_agent",
                    "research_agent",
                    "itinerary_agent",
                    "lodging_agent"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Kind"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "queued",
                    "running",
                    "succeeded",
                    "partial",
                    "failed",
                    "timed_out",
                    "cancelled",
                    "interrupted"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 30,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RunOut"
                  },
                  "title": "Response List Runs"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/runs/{run_id}": {
      "get": {
        "tags": [
          "automation"
        ],
        "summary": "Get Run",
        "operationId": "get_run",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunDetailOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/runs/{run_id}/events": {
      "get": {
        "tags": [
          "automation"
        ],
        "summary": "Run Events",
        "description": "Log lines after `after_seq`, for polling while a run is in progress.",
        "operationId": "run_events",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Run Id"
            }
          },
          {
            "name": "after_seq",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "After Seq"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RunEventOut"
                  },
                  "title": "Response Run Events"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/runs/{run_id}/outputs": {
      "get": {
        "tags": [
          "automation"
        ],
        "summary": "Run Outputs",
        "description": "What the run saved (prices, notes, suggested activities) and what was rejected, with the reasons.",
        "operationId": "run_outputs",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunOutputs"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/runs/{run_id}/cancel": {
      "post": {
        "tags": [
          "automation"
        ],
        "summary": "Cancel Run",
        "operationId": "cancel_run",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/routines": {
      "get": {
        "tags": [
          "automation"
        ],
        "summary": "List Routines",
        "operationId": "list_routines",
        "parameters": [
          {
            "name": "trip_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Trip Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RoutineOut"
                  },
                  "title": "Response List Routines"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "automation"
        ],
        "summary": "Add Routine",
        "operationId": "add_routine",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RoutineCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoutineOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/trips/{trip_id}/routines": {
      "get": {
        "tags": [
          "automation"
        ],
        "summary": "Trip Routines",
        "operationId": "trip_routines",
        "parameters": [
          {
            "name": "trip_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Trip Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RoutineOut"
                  },
                  "title": "Response Trip Routines"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/routines/{routine_id}": {
      "get": {
        "tags": [
          "automation"
        ],
        "summary": "Get Routine Detail",
        "operationId": "get_routine_detail",
        "parameters": [
          {
            "name": "routine_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Routine Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoutineOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "automation"
        ],
        "summary": "Update Routine",
        "operationId": "update_routine",
        "parameters": [
          {
            "name": "routine_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Routine Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RoutineUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoutineOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "automation"
        ],
        "summary": "Remove Routine",
        "operationId": "remove_routine",
        "parameters": [
          {
            "name": "routine_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Routine Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/routines/{routine_id}/run": {
      "post": {
        "tags": [
          "automation"
        ],
        "summary": "Run Routine Now",
        "description": "Queue a run now. If one is already waiting to start, that run is returned instead.",
        "operationId": "run_routine_now",
        "parameters": [
          {
            "name": "routine_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Routine Id"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/trips/{trip_id}/notes": {
      "get": {
        "tags": [
          "automation"
        ],
        "summary": "Trip Notes",
        "description": "Findings saved by research agents, newest first.",
        "operationId": "trip_notes",
        "parameters": [
          {
            "name": "trip_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Trip Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/NoteOut"
                  },
                  "title": "Response Trip Notes"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/notes/{note_id}": {
      "delete": {
        "tags": [
          "automation"
        ],
        "summary": "Delete Note",
        "operationId": "delete_note",
        "parameters": [
          {
            "name": "note_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Note Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "RunEventOut": {
        "properties": {
          "seq": {
            "type": "integer",
            "title": "Seq"
          },
          "ts": {
            "type": "string",
            "format": "date-time",
            "title": "Ts"
          },
          "type": {
            "type": "string",
            "title": "Type"
          },
          "tool_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tool Name"
          },
          "summary": {
            "type": "string",
            "title": "Summary"
          },
          "payload": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Payload"
          }
        },
        "type": "object",
        "required": [
          "seq",
          "ts",
          "type",
          "tool_name",
          "summary",
          "payload"
        ],
        "title": "RunEventOut"
      },
      "RunOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "routine_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Routine Id"
          },
          "trip_id": {
            "type": "integer",
            "title": "Trip Id"
          },
          "kind": {
            "type": "string",
            "enum": [
              "flight_api",
              "flight_agent",
              "research_agent",
              "itinerary_agent",
              "lodging_agent"
            ],
            "title": "Kind"
          },
          "trigger": {
            "type": "string",
            "enum": [
              "schedule",
              "manual",
              "catch_up"
            ],
            "title": "Trigger"
          },
          "status": {
            "type": "string",
            "enum": [
              "queued",
              "running",
              "succeeded",
              "partial",
              "failed",
              "timed_out",
              "cancelled",
              "interrupted"
            ],
            "title": "Status"
          },
          "params": {
            "additionalProperties": true,
            "type": "object",
            "title": "Params"
          },
          "queued_at": {
            "type": "string",
            "format": "date-time",
            "title": "Queued At"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "finished_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Summary"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "accepted_count": {
            "type": "integer",
            "title": "Accepted Count"
          },
          "rejected_count": {
            "type": "integer",
            "title": "Rejected Count"
          },
          "input_tokens": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Input Tokens"
          },
          "output_tokens": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Output Tokens"
          },
          "cost_usd_est": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cost Usd Est"
          },
          "cancel_requested": {
            "type": "boolean",
            "title": "Cancel Requested"
          }
        },
        "type": "object",
        "required": [
          "id",
          "routine_id",
          "trip_id",
          "kind",
          "trigger",
          "status",
          "params",
          "queued_at",
          "started_at",
          "finished_at",
          "summary",
          "error",
          "accepted_count",
          "rejected_count",
          "input_tokens",
          "output_tokens",
          "cost_usd_est",
          "cancel_requested"
        ],
        "title": "RunOut"
      },
      "RunOutputs": {
        "properties": {
          "quotes": {
            "items": {
              "$ref": "#/components/schemas/QuoteOut"
            },
            "type": "array",
            "title": "Quotes"
          },
          "notes": {
            "items": {
              "$ref": "#/components/schemas/NoteOut"
            },
            "type": "array",
            "title": "Notes"
          },
          "suggestions": {
            "items": {
              "$ref": "#/components/schemas/SuggestionOut"
            },
            "type": "array",
            "title": "Suggestions"
          },
          "rejections": {
            "items": {
              "$ref": "#/components/schemas/RejectionOut"
            },
            "type": "array",
            "title": "Rejections"
          }
        },
        "type": "object",
        "required": [
          "quotes",
          "notes",
          "suggestions",
          "rejections"
        ],
        "title": "RunOutputs",
        "description": "Everything a run saved or had rejected."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}